+++ src/main/java/com/gotcha/server/auth/dto/response/GoogleTokenResponse.java
@@ -1,6 +1,6 @@
-package com.gotcha.server.auth.dto;
+package com.gotcha.server.auth.dto.response;
 
-import com.gotcha.server.member.dto.LoginResponse;
+import com.gotcha.server.member.dto.response.LoginResponse;
 
 public record GoogleTokenResponse(
         String access_token, Integer expires_in, String refresh_token, String scope, String token_type, String id_token) {
