+++ src/main/java/com/gotcha/server/auth/dto/response/RefreshTokenResponse.java
@@ -1,4 +1,4 @@
-package com.gotcha.server.auth.dto;
+package com.gotcha.server.auth.dto.response;
 
 public record RefreshTokenResponse(String access_token, int expires_in) {
