+++ src/main/java/com/gotcha/server/auth/dto/response/TokenInfoResponse.java
@@ -1,4 +1,4 @@
-package com.gotcha.server.auth.dto;
+package com.gotcha.server.auth.dto.response;
 
 public record TokenInfoResponse(String sub, String email) {
     public String socialId() {
