+++ src/main/java/com/gotcha/server/auth/oauth/GoogleOAuth.java
@@ -1,9 +1,9 @@
 package com.gotcha.server.auth.oauth;
 
-import com.gotcha.server.auth.dto.GoogleTokenResponse;
-import com.gotcha.server.auth.dto.GoogleUserResponse;
-import com.gotcha.server.auth.dto.RefreshTokenResponse;
-import com.gotcha.server.auth.dto.TokenInfoResponse;
+import com.gotcha.server.auth.dto.response.GoogleTokenResponse;
+import com.gotcha.server.auth.dto.response.GoogleUserResponse;
+import com.gotcha.server.auth.dto.response.RefreshTokenResponse;
+import com.gotcha.server.auth.dto.response.TokenInfoResponse;
 import com.gotcha.server.auth.security.MemberDetailsService;
 import com.gotcha.server.global.exception.AppException;
 import com.gotcha.server.global.exception.ErrorCode;
