+++ src/main/java/com/gotcha/server/member/dto/request/RefreshTokenRequest.java
@@ -1,4 +1,4 @@
-package com.gotcha.server.member.dto;
+package com.gotcha.server.member.dto.request;
 
 public record RefreshTokenRequest(Long userId) {
