+++ src/main/java/com/gotcha/server/member/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.gotcha.server.member.dto;
+package com.gotcha.server.member.dto.response;
 
 public record LoginResponse(Long userId, String accessToken, String refreshToken) {
