+++ src/main/java/com/gotcha/server/member/service/MemberService.java
@@ -1,16 +1,15 @@
 package com.gotcha.server.member.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.gotcha.server.auth.dto.RefreshTokenResponse;
+import com.gotcha.server.auth.dto.response.RefreshTokenResponse;
 import com.gotcha.server.auth.oauth.GoogleOAuth;
 import com.gotcha.server.global.exception.AppException;
 import com.gotcha.server.global.exception.ErrorCode;
 import com.gotcha.server.member.domain.Member;
 import com.gotcha.server.member.domain.MemberRepository;
-import com.gotcha.server.auth.dto.GoogleTokenResponse;
-import com.gotcha.server.auth.dto.GoogleUserResponse;
-import com.gotcha.server.member.dto.LoginResponse;
-import com.gotcha.server.member.dto.RefreshTokenRequest;
+import com.gotcha.server.auth.dto.response.GoogleTokenResponse;
+import com.gotcha.server.auth.dto.response.GoogleUserResponse;
+import com.gotcha.server.member.dto.response.LoginResponse;
+import com.gotcha.server.member.dto.request.RefreshTokenRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -26,7 +25,7 @@ public class MemberService {
         return googleOAuth.getLoginUrl();
     }
 
-    public RefreshTokenResponse refresh(RefreshTokenRequest request) throws JsonProcessingException {
+    public RefreshTokenResponse refresh(RefreshTokenRequest request) {
         Member member = memberRepository.findById(request.userId())
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
         return googleOAuth.requestRefresh(member.getRefreshToken());
